test(qa): split Mantis process runtime coverage
